Better handling of snake buffer batch closing
What this does is that it introduces a command called SnakeClose
to close the additional windows that were opened by this command.
It always leaves the original window intact, unless SnakeClose is
given a ! to force it.

Likewise, SnakeBuffer was also modified to allow for a ! modifier
to change the number of columns forcefully associated with the snake.